# backend/app/schemas/chat_schemas.py
"""
AI Chatbot schemas for FreshMart conversational assistant.
"""

from datetime import datetime
from typing import Optional, List
from pydantic import BaseModel, Field

class ChatMessageRequest(BaseModel):
    """User conversational inquiry submitted to the AI Chatbot."""
    message: str = Field(
        ...,
        min_length=1,
        max_length=2000,
        description="Natural-language conversational message or inquiry.",
        example="What business domains am I authorized to access?"
    )
    conversation_id: Optional[str] = Field(
        default=None,
        description="Optional conversation thread identifier for session continuity.",
        example="chat-session-001"
    )

class ChatMessageResponse(BaseModel):
    """AI chatbot conversational response with authenticated user context."""
    reply: str = Field(..., description="Assistant conversational response generated by local LLM (Llama 3.1 8B)")
    conversation_id: Optional[str] = Field(None, description="Active conversation session identifier", example="chat-session-001")
    user_role: str = Field(..., description="Application role of the authenticated user", example="SALES_MANAGER")
    user_name: str = Field(..., description="Full name of the requesting user", example="Arun Kumar")
    allowed_domains: List[str] = Field(..., description="Business domains the user is authorized to query based on RBAC permissions", example=["CRM (Customers & Leads)", "Sales & Revenue", "ERP & Inventory"])
    model_used: str = Field(..., description="Ollama LLM model identifier utilized for generation", example="llama3.1:8b")
    timestamp: datetime = Field(default_factory=datetime.utcnow, description="UTC timestamp of reply generation")