# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FreshMart Agentic Analytics API
  description: "\n# FreshMart Agentic Analytics API\n\nWelcome to the **FreshMart\
    \ Agentic Analytics Platform** enterprise API documentation.\n\nFreshMart Agentic\
    \ Analytics provides secure, AI-powered natural-language analytics for FreshMart\
    \ enterprise business data across retail operations, customer relationship management,\
    \ human resources, inventory warehousing, purchasing, and financial performance.\n\
    \n---\n\n### Core Platform Capabilities\n\n* **AI-Powered Natural-Language Analytics:**\
    \ Inquire in plain English (e.g., *\"Why did profit decrease in August?\"* or\
    \ *\"Show monthly sales trend for the last 12 months\"*).\n* **LangGraph Multi-Agent\
    \ Orchestration:** Specialized agents collaborate autonomously:\n  * **Intent\
    \ Classifier:** Detects analytical goal, business domain, and inquiry category.\n\
    \  * **Permission Checker:** Enforces database table and column-level RBAC prior\
    \ to query generation.\n  * **Schema Planner:** Dynamically inspects schemas using\
    \ the Model Context Protocol (MCP).\n  * **SQL Generator & Corrector:** Synthesizes\
    \ read-only SQL via local `qwen2.5-coder:7b`.\n  * **SQL Validator:** Rigorous\
    \ AST parser preventing non-SELECT, destructive, or unauthorized queries.\n  *\
    \ **SQL Executor:** Safely runs validated SQL through MCP with row bounding and\
...
    \ `admin` / `AdminPassword123!`\n"
  contact:
...
paths:
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Authenticate User & Generate JWT Access Token
      description: 'Authenticates a FreshMart user with username and password, returning
        a signed JWT Bearer access token along with the user profile, role, and granted
        RBAC permission codes.


        **Supported Content-Types:**

        - `application/json`: Standard REST API clients

        - `application/x-www-form-urlencoded`: Interactive Swagger UI Authorize dialog
        and HTML forms


        **Available Demo Accounts (Password: `<Role>Password123!` or `AdminPassword123!`):**

        - `ceo`: Full enterprise analytics access

        - `sales.manager`: Sales, CRM, and inventory analytics

        - `finance.manager`: Financial P&L, expenses, and profit analytics

        - `hr.manager`: HRMS directory and employee salary analytics

        - `inventory.manager`: Products, warehouse stock, and purchasing

        - `erp.manager`: Integrated procurement, stock, sales, and HR

        - `admin`: System administration (user and role management)'
      operationId: loginUser
      requestBody:
        description: FreshMart application credentials submitted as JSON or Form URL-encoded
          data.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              ceo_login:
                summary: CEO Account (Full Access)
                description: Login as CEO with full enterprise-wide visibility.
                value:
                  username: ceo
                  password: CeoPassword123!
              sales_manager_login:
                summary: Sales Manager Account
                description: Login as Sales Manager with sales, CRM, and inventory
                  permissions.
                value:
                  username: sales.manager
                  password: SalesPassword123!
              finance_manager_login:
                summary: Finance Manager Account
                description: Login as Finance Manager with P&L, profit, and expense
                  access.
                value:
                  username: finance.manager
                  password: FinancePassword123!
              hr_manager_login:
                summary: HR Manager Account
                description: Login as HR Manager with employee salary inspection rights.
                value:
                  username: hr.manager
                  password: HrPassword123!
              admin_login:
                summary: System Administrator
                description: Login as Technical Administrator for user and role management.
                value:
                  username: admin
                  password: AdminPassword123!
          application/x-www-form-urlencoded:
            schema:
              properties:
                username:
                  type: string
                  description: Application username (e.g., ceo, sales.manager, admin)
                  example: ceo
                password:
                  type: string
                  format: password
                  description: Application password
                  example: CeoPassword123!
                grant_type:
                  type: string
                  default: password
              type: object
              required:
              - username
              - password
        required: true
      responses:
        '200':
          description: User successfully authenticated. JWT access token and user
            profile returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: 'Bad Request: Missing username or password, or malformed request
            payload.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '401':
          description: 'Unauthorized: Incorrect username or password.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User account is inactive or disabled by system
            administrator.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Request payload could not be processed.'
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get Current Authenticated User Profile
      description: Returns profile information, role metadata, and granted RBAC permission
        codes for the currently authenticated user identified by the verified JWT
        access token.
      operationId: getCurrentUserProfile
      responses:
        '200':
          description: User profile and granted permissions successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: 'Forbidden: User account is inactive or disabled.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
  /api/v1/chat/message:
    post:
      tags:
      - AI Chatbot
      summary: Send Conversational Business Inquiry
      description: 'Role-scoped conversational AI chat endpoint powered by Ollama
        Llama 3.1 8B:

        1. Authenticates user identity and role from backend JWT.

        2. Injects authoritative system context with permitted business domains.

        3. Queries local Ollama LLM with role-scoped context.

        4. Returns AI response with role and permitted domain metadata.'
      operationId: sendChatMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
        required: true
      responses:
        '200':
          description: Conversational assistant response generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '401':
          description: 'Unauthorized: Missing, invalid, or expired JWT access token.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          description: 'Validation Error: Request message payload could not be validated.'
        '503':
          description: 'Service Unavailable: Local Ollama LLM engine is offline or
            unreachable.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      security:
      - bearerAuth: []
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    ChatMessageRequest:
      properties:
        message:
          type: string
          maxLength: 2000
          minLength: 1
          title: Message
          description: Natural-language conversational message or inquiry.
          example: What business domains am I authorized to access?
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
          description: Optional conversation thread identifier for session continuity.
          example: chat-session-001
      type: object
      required:
      - message
      title: ChatMessageRequest
      description: User conversational inquiry submitted to the AI Chatbot.
    ChatMessageResponse:
      properties:
        reply:
          type: string
          title: Reply
          description: Assistant conversational response generated by local LLM (Llama
            3.1 8B)
        conversation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Conversation Id
          description: Active conversation session identifier
          example: chat-session-001
        user_role:
          type: string
          title: User Role
          description: Application role of the authenticated user
          example: SALES_MANAGER
        user_name:
          type: string
          title: User Name
          description: Full name of the requesting user
          example: Arun Kumar
        allowed_domains:
          items:
            type: string
          type: array
          title: Allowed Domains
          description: Business domains the user is authorized to query based on RBAC
            permissions
          example:
          - CRM (Customers & Leads)
          - Sales & Revenue
          - ERP & Inventory
        model_used:
          type: string
          title: Model Used
          description: Ollama LLM model identifier utilized for generation
          example: llama3.1:8b
        timestamp:
          type: string
          format: date-time
          title: Timestamp
          description: UTC timestamp of reply generation
      type: object
      required:
      - reply
      - user_role
      - user_name
      - allowed_domains
      - model_used
      title: ChatMessageResponse
      description: AI chatbot conversational response with authenticated user context.
...
    HTTPError:
      properties:
        detail:
          type: string
          title: Detail
          description: Human-readable explanation of why the request failed.
          example: 'Access Denied: You do not have permission ''VIEW_EMPLOYEE_SALARY''.'
      type: object
      required:
      - detail
      title: HTTPError
      description: Standard RFC 7807-compatible error response detail.
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
          description: Signed JWT Bearer access token containing claims and permissions
        token_type:
          type: string
          title: Token Type
          description: Token protocol type
          default: bearer
          example: bearer
        user:
          $ref: '#/components/schemas/UserProfile'
          description: Authenticated user profile and permissions
      type: object
      required:
      - access_token
      - user
      title: TokenResponse
      description: JWT token response returned upon successful authentication.
...
    UserProfile:
      properties:
        user_id:
          type: integer
          title: User Id
          description: Unique integer user identifier
          example: 2
        username:
          type: string
          title: Username
          description: Unique application username
          example: ceo
        email:
          type: string
          title: Email
          description: Corporate email address
          example: ceo@freshmart.local
        full_name:
          type: string
          title: Full Name
          description: User's full name
          example: Vikram Malhotra
        role:
          type: string
          title: Role
          description: Assigned application role name
          example: CEO
        employee_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Employee Id
          description: Linked MySQL workforce employee ID
          example: E006
        permissions:
          items:
            type: string
          type: array
          title: Permissions
          description: List of granular RBAC permission codes granted to this user
          example:
          - VIEW_HR
          - VIEW_EMPLOYEE_SALARY
          - VIEW_CRM
          - VIEW_SALES
          - VIEW_INVENTORY
          - VIEW_PURCHASES
          - VIEW_FINANCE
          - VIEW_PROFIT
          - VIEW_EXPENSES
        is_active:
          type: boolean
          title: Is Active
          description: Account active status flag
          default: true
          example: true
      type: object
      required:
      - user_id
      - username
      - email
      - full_name
      - role
      - permissions
      title: UserProfile
      description: Current authenticated user profile and granted permissions.
...
    LoginRequest:
      description: User login request payload.
      examples:
      - password: CeoPassword123!
        username: ceo
      - password: SalesPassword123!
        username: sales.manager
      - password: HrPassword123!
        username: hr.manager
      - password: AdminPassword123!
        username: admin
      properties:
        username:
          description: Application username (e.g. ceo, sales.manager, admin)
          example: ceo
          title: Username
          type: string
        password:
          description: Application account password
          example: CeoPassword123!
          title: Password
          type: string
      required:
      - username
      - password
      title: LoginRequest
      type: object
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme. Enter token directly
        or as 'Bearer <token>'.
    OAuth2PasswordBearer:
      type: oauth2
      description: 'OAuth2 Password Flow: Authenticate using username and password'
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
...